+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 ## [Unreleased]
+- Add AWS SSO sign-in button so credentials no longer need pasting.
 - Hide initial password section when logged out.
 
 ## [2026-01-22]